feat(memory): wire the memory pillar into the pipeline
The derrick-memory crate (store, lesson index, tag retrieval) existed
but nothing constructed or queried it. It is now live end to end:

- Seeding: derrick init (greenfield + brownfield) initialises the
  memory store under .derrick/ and seeds the survey-index reference
  entry (D55); --dry-run prints what would be seeded.
- Retrieval: all three local dispatchers (generic host-CLI, local
  copilot, claude) inject up to LESSON_RETRIEVAL_LIMIT relevant
  lessons — tag-matched against the ticket id/title — as a delimited
  'Lessons from previous work in this repo' block before the
  roughneck wrap. Empty index adds zero tokens.
- Curation: on successful run completion the flow runner records a
  deterministic run lesson (run id, pipeline, step outcomes, batch and
  ticket ids) gated by D9's mechanical anchor rule; failures log WARN
  and never break the run.
- Roughneck adoption: derrick-flow and the dispatchers now call the
  compliance-gated estimate_savings API instead of the deprecated
  unconditional estimate, so recorded savings reflect what the model
  actually did.
- README: test count updated to the real figure (801 across 20
  crates).

+++ crates/derrick-memory/src/lib.rs
@@ -748,6 +748,7 @@ pub fn inject_lessons_into_prompt(prompt: &str, index: &LessonIndex, query: &str
 ///
 /// [`relevant`]: LessonIndex::relevant
 /// [`recent`]: LessonIndex::recent
+#[derive(Debug)]
 pub struct LessonIndex {
     lessons: Vec<Lesson>,
     /// tag (ticket id or section anchor) → indices into `lessons`
@@ -1895,7 +1896,7 @@ state:
         for i in 1..=8u32 {
             store
                 .append_lesson(&lesson(
-                    utc(2026, 5, i as i32),
+                    utc(2026, 5, i),
                     &format!("drk-{i} lesson number {i}"),
                 ))
                 .unwrap();
